+++ R/MiDAS_class.R
@@ -195,7 +195,7 @@ setMethod(
 #' Get omnibus groups from MiDAS object.
 #'
 #' @details For some experiments features can be naturally divided into groups
-#' (here called omnibus groups). For example, in \code{'hla_aa'} experiment
+#' (here called omnibus groups). For example, in \code{"hla_aa"} experiment
 #' features can be grouped by amino acid position (\code{"B_46_E"},
 #' \code{"B_46_A"}) can be grouped into \code{B_46} group). Such groups can be
 #' then used to perform omnibus test, see \code{\link{runMiDAS}} for more
@@ -236,7 +236,13 @@ setMethod(
     assert_that(is.string(experiment))
     experiment <- object[[experiment]]
     omnibus_groups <- if (isClass(experiment, "SummarizedExperiment")) {
-      metadata(experiment)$omnibus_groups
+      og <- metadata(experiment)$omnibus_groups
+      if (is.null(og)) {
+        NULL
+      } else {
+        el <- rownames(experiment)
+        filterListByElements(og, el)
+      }
     } else {
       NULL
     }
@@ -261,7 +267,31 @@ setMethod(
 #'   frequency data frame should be formatted.
 #'
 #' @return Data frame with features from selected experiment and their
-#'   corresponding frequencies.
+#'   corresponding frequencies. Column \code{"term"} hold features names, 
+#'   \code{"Counts"} hold number of feature occurrences, \code{"Freq"} hold 
+#'   feature frequencies. If argument \code{compare} is set to \code{TRUE}, 
+#'   further columns will hold frequencies in reference populations.
+#'   
+#' @examples 
+#' midas <- prepareMiDAS(
+#'   hla_calls = MiDAS_tut_HLA,
+#'   colData = MiDAS_tut_pheno,
+#'   experiment = "hla_alleles"
+#' )
+#' 
+#' # using default reference populations
+#' getFrequencies(midas, experiment = "hla_alleles", compare = TRUE)
+#' 
+#' # using customized set of reference populations
+#' getFrequencies(
+#'   object = midas, 
+#'   experiment = "hla_alleles", 
+#'   compare = TRUE,
+#'   ref_pop = list(
+#'     hla_alleles = c("USA NMDP Chinese", "USA NMDP European Caucasian")
+#'   ),
+#'   ref = list(hla_alleles = allele_frequencies)
+#' )
 #'
 #' @importFrom assertthat assert_that is.string
 #' @importFrom dplyr select
@@ -525,7 +555,6 @@ setMethod(
 
     mask <- unlist(omnibus_groups[groups]) # TODO what if some groups are missing due to smth
     object[[experiment]] <- se[mask, ]
-    S4Vectors::metadata(object[[experiment]])$omnibus_groups <- omnibus_groups[groups]
 
     return(object)
   })
@@ -601,7 +630,7 @@ setMethod(
     assert_that(
       validObject(object),
       is.string(aa_pos),
-      see_if(grepl("^[A-Z]+[0-9]*_[0-9]+$", aa_pos),
+      see_if(grepl("^[A-Z]+[0-9]*_-*[0-9]+$", aa_pos),
              msg = "amino acid position should be formatted like: A_9."
       )
     )
@@ -630,6 +659,10 @@ setMethod(
 #' Coerce MiDAS to Data Frame
 #'
 #' @method as.data.frame MiDAS
+#' 
+#' @return Data frame representation of MiDAS object. Consecutive columns hold
+#'   values of variables from MiDAS's experiments and colData. The metadata
+#'   associated with experiments is not preserved.
 #'
 #' @inheritParams base::as.data.frame
 #' @export
@@ -1056,11 +1089,20 @@ prepareMiDAS_hla_NK_ligands <- function(hla_calls, ...) {
     checkHlaCallsFormat(hla_calls)
   )
 
-  lib <- c(
-    "allele_HLA_Bw",
-    "allele_HLA-Bw_only_B",
-    "allele_HLA-C_C1-2"
-  )
+  # Bw6 is defined both in Bw and Bw HLA-B only dictionaries, here this unambiguity is removed
+  bw_with_A <-
+    system.file("extdata", "Match_allele_HLA_Bw.txt", package = "MiDAS")
+  bw_with_A <-
+    read.table(
+      file = bw_with_A,
+      header = TRUE,
+      sep = "\t",
+      stringsAsFactors = FALSE
+    )
+  mask <- bw_with_A$group != "Bw6"
+  bw_with_A <- bw_with_A[mask, ]
+  
+  lib <- list(bw_with_A, "allele_HLA-Bw_only_B", "allele_HLA-C_C1-2")
   hla_NK_ligands <- Reduce(
     f = function(...) left_join(..., by = "ID"),
     x = lapply(lib, hlaToVariable, hla_calls = hla_calls, na.value = 0)