+++ R/MiDAS_transformationFunctions.R
@@ -220,7 +220,9 @@ hlaToAAVariation <- function(hla_calls,
 #' @param nacols.rm Logical indicating if result columns that contain only
 #'   \code{NA} should be removed.
 #'
-#' @return Data frame of HLA variables.
+#' @return Data frame with variable number of columns. First column named 
+#'   \code{"ID"} corresponds to \code{"ID"} column in \code{hla_calls}, further
+#'   columns holds converted HLA variables.
 #'
 #' @examples
 #' hlaToVariable(MiDAS_tut_HLA, dictionary = "allele_HLA_supertype")
@@ -308,7 +310,7 @@ hlaToVariable <- function(hla_calls,
 #' @inheritParams checkHlaCallsFormat
 #' @inheritParams reduceAlleleResolution
 #'
-#' @return HLA calls reduced to specified resolution.
+#' @return HLA calls data frame reduced to specified resolution.
 #'
 #' @examples
 #' reduceHlaCalls(MiDAS_tut_HLA, resolution = 2)
@@ -333,7 +335,9 @@ reduceHlaCalls <- function(hla_calls, resolution = 4) {
 #'   be checked. This is useful if one wants to use \code{hlaCallsToCounts} with
 #'   input not adhering to HLA nomenclature standards. See examples.
 #'
-#' @return HLA allele counts data frame.
+#' @return HLA allele counts data frame. First column holds samples ID's, further 
+#'   columns, corresponding to specific alleles, give information on the number 
+#'   of their occurrences in each sample.
 #'
 #' @importFrom assertthat assert_that is.string
 #' @importFrom qdapTools mtabulate
@@ -380,7 +384,8 @@ hlaCallsToCounts <- function(hla_calls,
 #' @param ref Data frame giving reference allele frequencies. See
 #'   \code{\link{allele_frequencies}} for an example.
 #'
-#' @return Data frame containing HLA alleles and their corresponding frequencies.
+#' @return Data frame with each row holding HLA allele, it's count and 
+#'   frequency.
 #'
 #' @examples
 #' getHlaFrequencies(MiDAS_tut_HLA)
@@ -437,7 +442,8 @@ getHlaFrequencies <- function(hla_calls,
 #'
 #' @inheritParams checkKirCallsFormat
 #'
-#' @return Data frame containing KIR genes and their corresponding frequencies.
+#' @return Data frame with each row holding KIR gene, it's count and 
+#'   frequency.
 #'
 #' @examples
 #' getKIRFrequencies(MiDAS_tut_KIR)
@@ -472,7 +478,9 @@ getKIRFrequencies <- function(kir_calls) {
 #' @param aa_variation Amino acid variation data frame as returned by
 #'   \link{hlaToAAVariation}.
 #'
-#' @return Amino acid counts data frame.
+#' @return Amino acid counts data frame. First column holds samples ID's, 
+#'   further columns, corresponding to specific amino acid positions, give 
+#'   information on the number of their occurrences in each sample.
 #'
 #' @importFrom assertthat assert_that is.string
 #' @importFrom qdapTools mtabulate
@@ -521,8 +529,8 @@ aaVariationToCounts <- function(aa_variation) {
 #'
 #' @inheritParams aaVariationToCounts
 #'
-#' @return Data frame containing amino acid positions and their corresponding
-#'   frequencies.
+#' @return Data frame with each row holding specific amino acid position, it's 
+#'   count and frequency.
 #'
 #' @examples
 #' aa_variation <- hlaToAAVariation(MiDAS_tut_HLA)
@@ -755,8 +763,10 @@ kableResults <- function(results,
 #' @param na.value Vector of length one speciyfing value for variables with no
 #'   matching entry in \code{dictionary}. Default is to use \code{0}.
 #'
-#' @return Data frame of indicators for new variables, with \code{1} and
-#'   \code{0} signaling presence and  absence of a variable respectively.
+#' @return Data frame with variable number of columns. First column named 
+#'   \code{"ID"} corresponds to \code{"ID"} column in \code{counts}, further
+#'   columns hold indicators for converted variables. \code{1} and \code{0} 
+#'   code presence and absence of a variable respectively.
 #'
 #' @examples
 #' countsToVariables(MiDAS_tut_KIR, "kir_haplotypes")
@@ -852,8 +862,17 @@ countsToVariables <- function(counts,
 #' @inheritParams checkKirCallsFormat
 #' @param interactions_dict Path to HLA - KIR interactions dictionary.
 #'
-#' @return Data frame with presence-absence indicators for HLA - KIR
-#'   interactions.
+#' @return Data frame with variable number of columns. First column named 
+#'   \code{"ID"} corresponds to \code{"ID"} column in \code{counts}, further
+#'   columns hold indicators for HLA - KIR interactions. \code{1} and \code{0} 
+#'   code presence and absence of a variable respectively.
+#'   
+#' @examples
+#' getHlaKirInteractions(
+#'   hla_calls = MiDAS_tut_HLA, 
+#'   kir_calls = MiDAS_tut_KIR, 
+#'   interactions_dict = system.file("extdata", "Match_counts_hla_kir_interactions.txt", package = "MiDAS")
+#' )
 #'
 #' @importFrom assertthat assert_that is.string
 #' @importFrom dplyr left_join
@@ -991,11 +1010,12 @@ filterExperimentByFrequency.SummarizedExperiment <-
   mask <- rownames(experiment) %in% filtered_vars
   new_experiment <- experiment[mask, , drop = FALSE]
   
-  og <- S4Vectors::metadata(experiment)$omnibus_groups
-  if (! is.null(og)) {
-    new_og <- filterListByElements(list = og, elements = filtered_vars)
-    metadata(new_experiment)$omnibus_groups <- new_og
-  }
+  # omnibus groubs are static the proper filtering takes place in getOmnibusGroups
+  # og <- S4Vectors::metadata(experiment)$omnibus_groups
+  # if (! is.null(og)) {
+  #   new_og <- filterListByElements(list = og, elements = filtered_vars)
+  #   metadata(new_experiment)$omnibus_groups <- new_og
+  # }
   
   return(new_experiment)
 }
@@ -1014,7 +1034,8 @@ filterExperimentByFrequency.SummarizedExperiment <-
 #'   following columns. See \code{\link{getReferenceFrequencies}} for more
 #'   details.
 #'
-#' @return Data frame containing variables and their corresponding frequencies.
+#' @return Data frame with each row holding specific variable, it's count and 
+#'   frequency.
 #'
 #' @importFrom assertthat assert_that is.string see_if
 #' @importFrom SummarizedExperiment assay
@@ -1110,7 +1131,7 @@ getExperimentFrequencies.SummarizedExperiment <-
 #' Under \code{"dominant"} model homozygotes and heterozygotes are coded as
 #' \code{1}. In \code{"recessive"} model homozygotes are coded as \code{1} and
 #' other as \code{0}. In \code{"additive"} model homozygotes are coded as
-#' \code{2} and heterozygotes as \code{1}. In \code{"overdominance"} homozygotes
+#' \code{2} and heterozygotes as \code{1}. In \code{"overdominant"} homozygotes
 #' (both \code{0} and \code{2}) are coded as \code{0} and heterozygotes as \code{1}.
 #'
 #' @param experiment Matrix or SummarizedExperiment object.
@@ -1122,7 +1143,7 @@ getExperimentFrequencies.SummarizedExperiment <-
 #'
 applyInheritanceModel <-
   function(experiment,
-           inheritance_model = c("dominant", "recessive", "additive", "overdominance")) {
+           inheritance_model = c("dominant", "recessive", "additive", "overdominant")) {
     UseMethod("applyInheritanceModel", experiment)
   }
 
@@ -1130,7 +1151,7 @@ applyInheritanceModel <-
 #' @method applyInheritanceModel matrix
 #'
 applyInheritanceModel.matrix <- function(experiment,
-                                         inheritance_model =  c("dominant", "recessive", "additive", "overdominance")) {
+                                         inheritance_model =  c("dominant", "recessive", "additive", "overdominant")) {
   .classifyGte <- function(x, val) {
   # classifies vector as being greater than number
   # specificly for 1 we classify as dominat, and 2 as recessive
@@ -1140,7 +1161,7 @@ applyInheritanceModel.matrix <- function(experiment,
   }
   .classifyEq <- function(x, val) {
   # classify vector as being equal to number
-  # specificly for 1 we classify as overdominance
+  # specificly for 1 we classify as overdominant
     x <- x == val
     mode(x) <- "integer"
     x
@@ -1150,7 +1171,7 @@ applyInheritanceModel.matrix <- function(experiment,
     "additive" = experiment,
     "dominant" = .classifyGte(experiment, 1), # ifelse(x >= 1, 1, 0)
     "recessive" = .classifyGte(experiment, 2), # ifelse(x >= 2, 1, 0)
-    "overdominance" = .classifyEq(experiment, 1) # ifelse(x == 1, 1, 0)
+    "overdominant" = .classifyEq(experiment, 1) # ifelse(x == 1, 1, 0)
   )
 }
 
@@ -1158,7 +1179,7 @@ applyInheritanceModel.matrix <- function(experiment,
 #' @method applyInheritanceModel SummarizedExperiment
 #'
 applyInheritanceModel.SummarizedExperiment <- function(experiment,
-                                                       inheritance_model =  c("dominant", "recessive", "additive", "overdominance")) {
+                                                       inheritance_model =  c("dominant", "recessive", "additive", "overdominant")) {
   SummarizedExperiment::assay(experiment) <-
     applyInheritanceModel(SummarizedExperiment::assay(experiment), inheritance_model)
 
@@ -1223,11 +1244,12 @@ filterExperimentByVariables.matrix <- function(experiment, variables) {
 filterExperimentByVariables.SummarizedExperiment <- function(experiment, variables) {
   og <- S4Vectors::metadata(experiment)$omnibus_groups
   experiment <- experiment[variables, ]
-  ## check if og is not null
-  if (! is.null(og)) {
-    og <- filterListByElements(list = og, elements = variables)
-    S4Vectors::metadata(experiment)$omnibus_groups <- og
-  }
+  
+  # omnibus groubs are static the proper filtering takes place in getOmnibusGroups
+  # if (! is.null(og)) {
+  #   og <- filterListByElements(list = og, elements = variables)
+  #   S4Vectors::metadata(experiment)$omnibus_groups <- og
+  # }
 
   return(experiment)
 }